+++ Vinyl/Track.swift
@@ -40,48 +40,3 @@ extension Track {
         response = Response(encodedResponse: encodedResponse)
     }
 }
-
-extension NSURLRequest {
-    
-    class func createWithEncodedRequest(encodedRequest: EncodedObject) -> NSURLRequest {
-        guard
-            let urlString = encodedRequest["url"] as? String,
-            let url = NSURL(string: urlString)
-        else {
-            fatalError("URL not found 😞 for Request: \(encodedRequest)")
-        }
-        
-        let request = NSMutableURLRequest(URL: url)
-        
-        if let method = encodedRequest["method"] as? String {
-            request.HTTPMethod = method
-        }
-        
-        if let headers = encodedRequest["headers"] as? HTTPHeaders {
-            request.allHTTPHeaderFields = headers
-        }
-        
-        if let body = encodedRequest["body"] as? NSData {
-            request.HTTPBody = body
-        }
-        
-        return request
-    }
-}
-
-extension Response {
-    
-    init(encodedResponse: EncodedObject) {
-        guard
-            let urlString = encodedResponse["url"] as? String,
-            let url =  NSURL(string: urlString),
-            let statusCode = encodedResponse["statusCode"] as? Int,
-            let headers = encodedResponse["headers"] as? HTTPHeaders,
-            let urlResponse = NSHTTPURLResponse(URL: url, statusCode: statusCode, HTTPVersion: nil, headerFields: headers)
-        else {
-            fatalError("key not found 😞 for Response (check url/statusCode/headers) \(encodedResponse)")
-        }
-        
-        self.init(urlResponse: urlResponse, body: decodeBody(encodedResponse["body"], headers: headers), error: nil)
-    }
-}
